oddsportal refactor progress
