+++ modules/alerts/alerts_formatter/matchup_streak_alert.py
@@ -457,7 +457,7 @@ def send_matchup_streak_alerts(notifier: Any, streak_reports: List) -> bool:
 
     try:
         from infrastructure.settings import Config
-        from oddsportal_config import SEASON_ODDSPORTAL_MAP
+        from modules.oddsportal.oddsportal_config import SEASON_ODDSPORTAL_MAP
         from infrastructure.persistence.repositories import EventRepository
     except ImportError:
         Config = None
