+++ modules/jobs/pre_start_check_job/alert_pipeline.py
@@ -16,7 +16,7 @@ from modules.alerts.alerts_formatter.matchup_streak_alert import send_matchup_st
 from modules.alerts.alerts_formatter.odds_alert import send_odds_alert
 from modules.alerts.matchup_streak_analysis import build_matchup_streak_context, should_send_streak_alert
 from modules.prediction import prediction_logger
-from oddsportal_config import SEASON_ODDSPORTAL_MAP
+from modules.oddsportal.oddsportal_config import SEASON_ODDSPORTAL_MAP
 from modules.sofascore import api_client
 
 logger = logging.getLogger(__name__)
