+++ modules/oddsportal/models.py
@@ -1,0 +1,9 @@
+from .dataclasses import (
+    BookieOdds,
+    BetfairExchangeOdds,
+    MarketExtraction,
+    MatchOddsData,
+    ScrapeAttemptResult,
+    GroupSeedResult,
+)
+
