+++ modules/oddsportal/oddsportal_scraper_core.py
@@ -1,0 +1,8 @@
+class OddsPortalScraper:
+    """Lazy wrapper that instantiates the legacy root scraper on demand."""
+
+    def __new__(cls, *args, **kwargs):
+        from oddsportal_scraper import OddsPortalScraper as RootOddsPortalScraper
+
+        return RootOddsPortalScraper(*args, **kwargs)
+
